feat: renamed session arg to chat_session
