+++ src/albert/resources/design.py
@@ -1,5 +1,4 @@
 import sys
-from uuid import UUID
 
 from pydantic import Field
 
@@ -34,27 +33,6 @@ class DesignRunViolationCode(StrEnum):
     OPTIMIZATION_SYSTEM_MISMATCH = "optimization_system_mismatch"
     INTERNAL = "internal"
     JOB_TIMEOUT = "job_timeout"
-
-
-class ChatSessionRef(BaseAlbertModel):
-    """A chat session to notify when an asynchronous job finishes.
-
-    Both identifiers are required. They are issued by the chat platform and
-    are normally supplied automatically by the agent runtime; a script calling the
-    SDK directly has no reason to construct one.
-
-    Notes
-    -----
-    This is a general async-callback context rather than a design-run concept, so
-    the same object is expected to carry completion routing for other long-running
-    jobs in future releases.
-    """
-
-    source_session_id: UUID = Field(alias="sourceSessionId")
-    """The originating frontend session identifier."""
-
-    chat_session_id: str = Field(alias="chatSessionId")
-    """The chat session identifier (``SES…``) that receives the completion message."""
 
 
 class DesignRunViolation(BaseAlbertModel):
