+++ tests/collections/test_design_runs.py
@@ -10,7 +10,6 @@ from albert.resources.targets import (
     NumericRange,
 )
 
-pytestmark = pytest.mark.xdist_group("datatemplates")
 
 ignore_in_ten0 = pytest.mark.xfail(
     reason="No DWH available in TEN0 test environment.",
